[Policies] complete policies for existing actions

+++ app/controllers/child_presents_controller.rb
@@ -13,6 +13,8 @@ class ChildPresentsController < ApplicationController
   end
 
   def destroy
+    authorize ChildPresent
+
     @present = @child_profile.child_presents.find(params[:id])
 
     if @present.destroy
